+++ src/ARevillaSearchFight/Views/Models/TermSearchResult.cs
@@ -5,12 +5,11 @@ using System.Threading.Tasks;
 
 namespace ARevillaSearchFight.Views.Models
 {
-
     public class TermSearchResult
     {
         public string SearchEngineName { get; set; }
 
-        public int Count { get; set; }
+        public long Count { get; set; }
 
         public string Term { get; set; }
     }
